pyonedesk/requirements.txt
hurry.filesize
pytest
click
requests
sanic
diskcache
progress
pycryptodome
shortuuid
colorama
setuptools
wheel
appdirs